requirements.txt
aiofile      ~=3.7.1
ircchallenge ~=0.1.2
ircrobots    ~=0.3.13
pyyaml       ~=5.4.1
